Move epub and comic modules into sub-directories. Add support for epub:type attributes.
